move components to subdirs
